requirements.txt
python-dotenv~=1.0.1
pynacl
python-telegram-bot
argparse~=1.4.0
requests~=2.32.3
keyring~=25.2.1
fastapi~=0.111.0
InquirerPy~=0.3.4
uvicorn~=0.30.1
pydantic~=2.7.4